Move packages from internal to pkg
AKSv1 needs to reuse CheckLogAnalyticsWorkspaceForMonitoring, which contains
metadata around the region calls. For that all internal packages should be
moved from internal to pkg directory.
