+++ pkg/aks/delete.go
